requirements.txt
Flask>=2.3.0
flask-cors>=4.0.0
python-dotenv
gunicorn
requests
unidecode
pytest
black
mypy
snowflake-connector-python>=3.16.0
pandas
cryptography
